Manifest, Comments and updated MainEngine
Partially completed comments file is added, as well as a completed
manifest.mf file (unsure if this is correct or not).
Found some errors with the printing out of win/lose (did not detect when
someone had won by both methods). It also now detects when both players
have won, and by which method they have won.
Input 3 is for a completely filled board when black wins
input 4 is for a completely filled board when both players win