Changed the gameboard constructor and extra stuff
Gameboard constructor now creates a new list of hexagons. Also tripod
win now looks for the minimum number instead of the maximum number.
Added some extra things to deal with the fact that there may be multiple
ways to get to an edge but they are not complete yet. we can talk about
it in class