Think I have figured out the problem
The problem is that the tripod evaluation function gives an equal value
to all of the edge pieces. So for example, if we hard-code the AI to
make a move at (0,1) and then (1,1) we would expect the next move to be
at (1,0) but as it values all edge pieces the same it will go to the
first edge piece it finds (0,2).

Another thing I noticed is that all of the nodes aren't being checked
for the tripod eval... I will look at this for a  bit longer and then I
will go to bed :)